move inverse depth to their own files
