+++ src/matches_containers.hpp
@@ -1,6 +1,7 @@
 #ifndef RGBDSLAM_UTILS_MATCHESCONTAINERS_HPP
 #define RGBDSLAM_UTILS_MATCHESCONTAINERS_HPP
 
+#include "utils/coordinates/inverse_depth_coordinates.hpp"
 #include "utils/coordinates/point_coordinates.hpp"
 #include "utils/coordinates/plane_coordinates.hpp"
 #include <list>
